feat: add file upload

+++ src/application/controller/itemController.ts
@@ -1,5 +1,4 @@
 import { Request, Response } from "express";
-import { Item } from "@model/itemEntity";
 import {
   ok,
   serverError,
@@ -10,11 +9,13 @@ import {
 } from "@utils/httpErrors";
 import { NotFoundError, BadRequestError } from "@utils/errors";
 import { IItemController, IItemService } from "@interfaces/index";
-import { ItemDTO } from "@dto/ItemDTO";
-import { validate, ValidationError } from "class-validator";
+import { ItemDTO } from "@dto/item/ItemDTO";
+import { ItemUpdateDTO } from "@dto/index";
+import { ValidationError } from "class-validator";
+import fs from 'fs';
 
 export class ItemController implements IItemController {
-  constructor(private readonly itemService: IItemService) {}
+  constructor(private readonly itemService: IItemService) { }
 
   async getItem(_: Request, res: Response): Promise<void> {
     try {
@@ -45,19 +46,34 @@ export class ItemController implements IItemController {
 
   async createItem(req: Request, res: Response): Promise<void> {
     const itemDTO = Object.assign(new ItemDTO(), req.body);
-
-    const validationErrors = await this.validateItemDTO(itemDTO);
-    if (validationErrors.length > 0) {
-      return badRequest(
-        res,
-        new BadRequestError(this.formatValidationErrors(validationErrors)),
-      );
-    }
+    let fileUrl: string | null = null;
 
     try {
-      const item = await this.itemService.create(itemDTO);
-      created(res, item);
+      const validationErrors = await this.itemService.validateItemDTO(itemDTO);
+      if (validationErrors.length > 0) {
+        if (req.file) {
+          fs.unlinkSync(req.file.path);
+        }
+        return badRequest(
+          res,
+          new BadRequestError(this.formatValidationErrors(validationErrors))
+        );
+      }
+
+      if (req.file) {
+        fileUrl = await this.itemService.handleFileUpload(req.file);
+        if (!fileUrl) {
+          return badRequest(res, new BadRequestError('Erro ao processar o arquivo.'));
+        }
+        itemDTO.url = fileUrl;
+      }
+
+      await this.itemService.create(itemDTO);
+      created(res, itemDTO);
     } catch (error: any) {
+      if (fileUrl) {
+        await this.itemService.deleteFile(fileUrl);
+      }
       serverError(res, error);
     }
   }
@@ -65,15 +81,45 @@ export class ItemController implements IItemController {
   async updateItem(req: Request, res: Response): Promise<void> {
     const itemId = this.extractItemId(req.params.id);
 
-    if (itemId === null) {
+    if (!itemId) {
       return badRequest(res, new BadRequestError("Invalid item ID"));
     }
 
+    const updatedItemDTO = Object.assign(new ItemUpdateDTO(), req.body);
+    let fileUrl: string | null = null;
+
     try {
-      const updatedItem: Item = req.body;
-      await this.itemService.update(itemId, updatedItem);
+      const validationErrors = await this.itemService.validateItemDTO(updatedItemDTO);
+      if (validationErrors.length > 0) {
+        if (req.file) {
+          fs.unlinkSync(req.file.path);
+        }
+        return badRequest(res, new BadRequestError(this.formatValidationErrors(validationErrors)));
+      }
+
+      const existingItem = await this.itemService.findOne(itemId);
+      if (!existingItem) {
+        return notFound(res, new NotFoundError("Item not found"));
+      }
+
+      if (req.file) {
+        fileUrl = await this.itemService.handleFileUpload(req.file);
+        if (!fileUrl) {
+          return badRequest(res, new BadRequestError('Erro ao processar o arquivo.'));
+        }
+
+        if (existingItem.url) {
+          await this.itemService.deleteFile(existingItem.url);
+        }
+        updatedItemDTO.url = fileUrl;
+      }
+
+      await this.itemService.update(itemId, updatedItemDTO);
       noContent(res);
     } catch (error: any) {
+      if (fileUrl) {
+        await this.itemService.deleteFile(fileUrl);
+      }
       serverError(res, error);
     }
   }
@@ -86,6 +132,15 @@ export class ItemController implements IItemController {
     }
 
     try {
+      const item = await this.itemService.findOne(itemId);
+      if (!item) {
+        return notFound(res, new NotFoundError("Item not found"));
+      }
+
+      if (item.url) {
+        await this.itemService.deleteFile(item.url);
+      }
+
       await this.itemService.delete(itemId);
       noContent(res);
     } catch (error: any) {
@@ -93,14 +148,10 @@ export class ItemController implements IItemController {
     }
   }
 
-  /* Other methods */
+  /* Helper methods */
   private extractItemId(id: string): number | null {
     const itemId = parseInt(id, 10);
     return isNaN(itemId) ? null : itemId;
-  }
-
-  private async validateItemDTO(itemDTO: ItemDTO): Promise<ValidationError[]> {
-    return await validate(itemDTO);
   }
 
   private formatValidationErrors(errors: ValidationError[]): string {
